Add functionality to View Cart
The user should be able to see the details of a cart
on a separate route. Also, there should be a way to put
a link on the home page to take user to cart details page.
The link icon should also show the quantity of items in the
cart.

- Added controller and views to show the cart items.
- Added icon and link for cart details in the nav bar.